requirements.txt
# Core dependencies
graphiti-core
python-dotenv

# Terminal output formatting for beautiful demo display
rich
colorama
